Resolves #58: Moving `printf.h` and `printf.c` into the `src/` directory (and marking it as an include directory).
